functions/etl-worker/requirements.txt
functions-framework==3.*
google-cloud-bigquery>=3.0.0
google-cloud-pubsub>=2.0.0
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0
